+++ CHANGES.md
@@ -6,6 +6,7 @@ Release Notes.
 ------------------
 #### Features
 * Add the compat protocol receiver for the old version of agents.
+* Support transmit the native eBPF Process and Profiling protocol.
 
 #### Bug Fixes
 * Fix Metadata lost in the Native Meter protocol.